+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Version 2.0.0
   Clusterio core.  Previously this was handled by the playerManager plugin.
 - Added stripping of long paths in the Factorio server log.
 - Added option to configure maximum number of commands to send in parallel.
+- Added option to auto start instances on slave startup.
 
 ### Changes